+++ src/pagerankComponentwiseCuda.hxx
@@ -25,7 +25,7 @@ using std::swap;
 // -------------
 
 template <class T, class J>
-int pagerankComponentwiseCudaLoop(T *e, T *r0, T *eD, T *r0D, T *&aD, T *&rD, T *cD, const T *fD, const int *vfromD, const int *efromD, int i, const J& ws, int N, T p, T E, int L, int EF) {
+int pagerankComponentwiseCudaLoop(T *e, T *r0, T *eD, T *r0D, T *&aD, T *&rD, T *cD, const T *fD, const size_t *vfromD, const int *efromD, int i, const J& ws, int N, T p, T E, int L, int EF) {
   float l = 0;
   for (const auto& w : ws) {
     const auto& [nt, nb] = w; int n = nt+nb;
